% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.gc.cont.r
\name{count.gc.cont}
\alias{count.gc.cont}
\title{A function to count GC content in the selected bins.}
\usage{
count.gc.cont(gcc = NULL, chrn = NULL, bin = 300, mc.cores = 2)
}
\arguments{
\item{gcc}{a vector containing information about GC content.}

\item{chrn}{a vector of chromosome names or NULL for all chromosomes.}

\item{bin}{a size of a bin.}

\item{mc.cores}{a number of cores for parallel computing.}
}
\description{
A function to count GC content in the selected bins.
}
\examples{
count.gc.cont()
}
\keyword{GC}
\keyword{content}

