% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.tags.in.bins.r
\name{count.tags.in.bins}
\alias{count.tags.in.bins}
\title{A function to compute number of tags in the selected bins.}
\usage{
count.tags.in.bins(profs = NULL, chr.stat = NULL, chrn = NULL,
  mc.cores = 2, bin = NULL, scale.genome.to.100Mb = TRUE,
  normalize.perBinSize = TRUE)
}
\arguments{
\item{profs}{a list of signle positions representing used reads.}

\item{chr.stat}{a data.frame with sizes of analyzed chromosmes.}

\item{chrn}{a vector of chromosome names or NULL for all chromosomes.}

\item{mc.cores}{a number of cores for parallel computing.}

\item{bin}{a size of a bin.}

\item{scale.genome.to.100Mb}{logical. Should the values be normalized to genome size?}

\item{normalize.perBinSize}{logical. Should the values be normalized to been size?}
}
\description{
A function to compute number of tags in the selected bins.
}
\examples{
count.tags.in.bins()
}
\keyword{bins}
\keyword{counts}

