% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.profiles.r
\name{generate.profiles}
\alias{generate.profiles}
\title{A function to read positions of aligned tags.}
\usage{
generate.profiles(tags, chrn = chrn, mc.cores = 2)
}
\arguments{
\item{tags}{a list of signle positions representing used reads.}

\item{chrn}{a vector of chromosome names or NULL for all chromosomes.}

\item{mc.cores}{a number of cores for parallel computing.}
}
\description{
A function to read positions of aligned tags.
}
\examples{
generate.profiles()
}
\keyword{tags}

