\name{macc-package}
\alias{macc-package}
\docType{package}
\title{
Causal Mediation Analysis under Correlated Errors
}
\description{
macc performs causal mediation analysis under confounding or correlated errors. This package includes a single level mediation model, a two-level mediation model and a three-level mediation model for data with hierarchical structure. Under the two/three-level mediation model, the correlation parameter is identifiable and estimated based on a hierarchical-likelihood or a two-stage method.
}
\details{
\tabular{ll}{
Package: \tab macc \cr
Type: \tab Package \cr
Version: \tab 1.0.1 \cr
Date: \tab 2017-08-20 \cr
License: \tab GPL (>=2) \cr
}
}
\author{
	Yi Zhao <yi_zhao@alumni.brown.edu> and Xi Luo <xi.rossi.luo@gmail.com> \cr

Maintainer: Yi Zhao <yi_zhao@alumni.brown.edu>
}
\references{Zhao, Y., & Luo, X. (2014). \emph{Estimating Mediation Effects under Correlated Errors with an Application to fMRI.} arXiv preprint arXiv:1410.7217.
}
\keyword{ package }
