% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFragmentLengthDensities.r
\name{plotFragmentLengthDensities}
\alias{plotFragmentLengthDensities}
\title{A function to plot distributions of fragment lengths.}
\usage{
plotFragmentLengthDensities(lengths = ll, col = c("#CDCD00", "#A0B43C",
  "#739B78", "#4682B4"), xlim = NULL)
}
\arguments{
\item{lengths}{a list of tag lengths. This may be retuned by read.tags function.}

\item{col}{a vector of colors.}

\item{xlim}{a numeric vector of length 2 giving range for x-axis.}
}
\description{
A function to plot distributions of fragment lengths.
}
\examples{
plotFragmentLengthDensities()
}
\keyword{lengths}

