% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGCvsMACC.r
\name{plotGCvsMACC}
\alias{plotGCvsMACC}
\title{A function to plot scatter plot showing association between MACC scores and GC content}
\usage{
plotGCvsMACC(data, sname, ylim = NULL, gc.cont = gc.cont)
}
\arguments{
\item{data}{a table which typically is one of the element of a list return by macc function.}

\item{sname}{a character string giving the header of the plot.}

\item{ylim}{a vector containing ranges for the y-axis.}

\item{gc.cont}{a results of ount.gc.count function containing GC content in the selected bins.}
}
\description{
A function to plot scatter plot showing association between MACC scores and GC content
}
\examples{
plotGCvsMACC()
}
\keyword{GC}
\keyword{correction}

