% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile2bedGraph.r
\name{profile2bedGraph}
\alias{profile2bedGraph}
\title{A wrapper function to compute MNase accessibility (MACC) scores with one command.}
\usage{
profile2bedGraph(tab, path = "./", fname = "profile", nn = 4)
}
\arguments{
\item{tab}{a matrix with four columns corresponding to chromosome names, start of the bins, end of the bins and interesting values respectively.}

\item{fname}{a character string giving the name of the file to be created.}

\item{nn}{integer indicating the number of decimal places to be used.}

\item{path2files}{a character string giving the name of directory where the new file will be saved.}
}
\description{
A wrapper function to compute MNase accessibility (MACC) scores with one command.
}
\examples{
profile2bedGraph()
}
\keyword{bedGraph}

