% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bam.file.r
\name{read.bam.file}
\alias{read.bam.file}
\title{A function to count GC content in the selected bins.}
\usage{
read.bam.file(fname, mc.cores = 2, tl = 50)
}
\arguments{
\item{fname}{a character showing a path to files to be read.}

\item{mc.cores}{a number of cores for parallel computing.}

\item{tl}{an integer for tag shifting.}
}
\description{
A function to count GC content in the selected bins.
}
\examples{
read.bam.file()
}
\keyword{bam}

