% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rta.r
\name{rta}
\alias{rta}
\title{A function to remove tag anomalies.}
\usage{
rta(tv)
}
\arguments{
\item{tv}{a vector with tag coordinates from the same strand.}
}
\description{
A function to remove tag anomalies.
}
\examples{
rta()
}
\keyword{tags}

