% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.centers.r
\name{tag.centers}
\alias{tag.centers}
\title{A function to compute positions of sequenced tags.}
\usage{
tag.centers(cp = NULL, max.length = 160, shift = 73)
}
\arguments{
\item{cp}{a vector with tag coordinates.}

\item{max.length}{an integer corresponding to maximal tag length.}

\item{shift}{an integer corresponding to shift of location applied to tags with length longer than max.length.}
}
\description{
A function to compute positions of sequenced tags.
}
\examples{
tag.centers()
}
\keyword{tags}

