% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.lengths.r
\name{tag.lengths}
\alias{tag.lengths}
\title{A function to compute lengths of sequenced tags.}
\usage{
tag.lengths(cp = NULL)
}
\arguments{
\item{cp}{a vector with tag coordinates.}
}
\description{
A function to compute lengths of sequenced tags.
}
\examples{
tag.lengths()
}
\keyword{tags}

