% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macc.r
\name{macc}
\alias{macc}
\title{A function to compute MNase accessibility (MACC) scores.}
\usage{
macc(tags.in.bins = NULL, tit.points = NULL, gc.cont = NULL,
  chrn = NULL, chr.stat = NULL, mc.cores = 2, bin = 300, CpG = NULL)
}
\arguments{
\item{tags.in.bins}{an list returned by count.tags.in.bins function and contatining localization of bins and number of tags in each of the selected bins.}

\item{tit.points}{a vector of MNase concentrations used in the experiment. The concentrations should be provided in ascending order.}

\item{gc.cont}{a list of vectors containing information about GC content. Each vector corresponds to different chromosome.}

\item{chrn}{a vector of chromosome names or NULL for all chromosomes.}

\item{chr.stat}{a data.frame with sizes of analyzed chromosmes.}

\item{mc.cores}{a number of cores for parallel computing.}

\item{bin}{a size of a bin.}

\item{CpG}{a GenomeRanges object with localization of CpG island.}

\item{scale.genome.to.100Mb}{logical. Should the values be normalized to genome size?}

\item{normalize.perBinSize}{logical. Should the values be normalized to been size?#' @keywords MACC}
}
\description{
A function to compute MNase accessibility (MACC) scores.
}
\examples{
macc()
}

